Move models in individual files
