+++ src/endpoints/language.js
@@ -1,19 +1,5 @@
 import Endpoint from './Endpoint'
-
-export class LanguageModel {
-  constructor (code, name) {
-    this._code = code
-    this._name = name
-  }
-
-  get code () {
-    return this._code
-  }
-
-  get name () {
-    return this._name
-  }
-}
+import LanguageModel from './models/LanguageModel'
 
 export default new Endpoint(
   'languages',
