+++ src/endpoints/location.js
@@ -1,30 +1,6 @@
 import { groupBy, sortBy } from 'lodash/collection'
-import { isEmpty } from 'lodash/lang'
 import Endpoint from './Endpoint'
-
-export class LocationModel {
-  constructor (name, path, live) {
-    this._code = name
-    this._name = path
-    this._live = live
-  }
-
-  get live () {
-    return this._live
-  }
-
-  get name () {
-    return this._code
-  }
-
-  get path () {
-    return this._name
-  }
-
-  get category () {
-    return isEmpty(this._code) ? '?' : this._code[0].toUpperCase()
-  }
-}
+import LocationModel from './models/LocationModel'
 
 export default new Endpoint(
   'locations',
