+++ src/endpoints/page.js
@@ -1,6 +1,7 @@
 import { transform } from 'lodash/object'
 import { forEach } from 'lodash/collection'
 import Endpoint from './Endpoint'
+import PageModel from './models/PageModel'
 
 export default new Endpoint(
   'pages',
@@ -38,44 +39,5 @@ export default new Endpoint(
     return new PageModel(0, options.location, 0, '', null, children)
   }
 )
-
-export class PageModel {
-  constructor (id = -1, title = '', parent = 0, content = '', thumbnail = null, children = {}) {
-    this._id = id
-    this._title = title
-    this._content = content
-    this._children = children
-    this._parent = parent
-    this._thumbnail = thumbnail
-  }
-
-  get thumbnail () {
-    return this._thumbnail
-  }
-
-  addChild (page) {
-    this._children[page.id] = page
-  }
-
-  get id () {
-    return this._id
-  }
-
-  get title () {
-    return this._title
-  }
-
-  get content () {
-    return this._content
-  }
-
-  get parent () {
-    return this._parent
-  }
-
-  get children () {
-    return this._children
-  }
-}
 
 export const EMPTY_PAGE = new PageModel()
